+++ src/osm_easy_api/api/endpoints/changeset.py
@@ -114,7 +114,7 @@ class Changeset_Container:
         match status_code:
             case 200: pass
             case 404: raise exceptions.IdNotFoundError()
-            case _: assert False, f"Unexpected response status code {status_code}. Please report it on github."
+            case _: assert False, f"Unexpected response status code {status_code}. Please report it on github." # pragma: no cover
 
         return self._xml_to_changeset(generator, include_discussion)[0] # type: ignore
 
@@ -169,7 +169,7 @@ class Changeset_Container:
             case 200: pass
             case 400: raise ValueError("Invalid arguments. See https://wiki.openstreetmap.org/wiki/API_v0.6#Query:_GET_/api/0.6/changesets for more info.")
             case 404: raise exceptions.IdNotFoundError()
-            case _: assert False, f"Unexpected response status code {status_code}. Please report it on github."
+            case _: assert False, f"Unexpected response status code {status_code}. Please report it on github." # pragma: no cover
 
         return self._xml_to_changeset(generator) # type: ignore
     
@@ -217,7 +217,7 @@ class Changeset_Container:
             case 200: pass
             case 404: raise exceptions.IdNotFoundError()
             case 409: raise exceptions.ChangesetAlreadyClosedOrUserIsNotAnAuthor(response.text)
-            case _: assert False, f"Unexpected response status code {response.status_code}. Please report it on github."
+            case _: assert False, f"Unexpected response status code {response.status_code}. Please report it on github." # pragma: no cover
 
         response.raw.decode_content = True
         return self._xml_to_changeset(self.outer._raw_stream_parser(response.raw), True)[0]
@@ -237,7 +237,7 @@ class Changeset_Container:
             case 200: pass
             case 404: raise exceptions.IdNotFoundError()
             case 409: raise exceptions.ChangesetAlreadyClosedOrUserIsNotAnAuthor(response.text)
-            case _: assert False, f"Unexpected response status code {response.status_code}. Please report it on github."
+            case _: assert False, f"Unexpected response status code {response.status_code}. Please report it on github." # pragma: no cover
 
     def download(self, id: int) -> Generator[Tuple['Action', 'Node | Way | Relation'], None, None]:
         """Download changes made in changeset. Like in 'diff' module.
@@ -256,7 +256,7 @@ class Changeset_Container:
         match stream.status_code:
             case 200: pass
             case 404: raise exceptions.IdNotFoundError()
-            case _: assert False, f"Unexpected response status code {stream.status_code}. Please report it on github."
+            case _: assert False, f"Unexpected response status code {stream.status_code}. Please report it on github." # pragma: no cover
         
         stream.raw.decode_content = True
         def generator() -> Generator[tuple['Action', 'Node | Way | Relation'], None, None]: